+++ src/CascLib.h
@@ -175,7 +175,7 @@ typedef enum _CASC_FILE_INFO_CLASS
 typedef struct _QUERY_KEY
 {
     LPBYTE pbData;
-    DWORD cbData;
+    size_t cbData;
 } QUERY_KEY, *PQUERY_KEY;
 
 // Query size for a file. Contains CSize + ESize
